+++ packages/vorbis_rs/src/encoder.rs
@@ -2,4 +2,4 @@ mod encoder_util;
 pub use encoder_util::{VorbisBitrateManagementStrategy, VorbisEncodingState};
 
 mod encoder_impl;
-pub use encoder_impl::VorbisEncoder;
+pub use encoder_impl::{VorbisEncoder, VorbisEncoderBuilder};